Modified the rx script to automatically select ibverbs device from ip address, cleaned up some of the code.
